+++ pilot/pkg/networking/core/v1alpha3/route/route_cache_test.go
@@ -81,3 +81,47 @@ func TestClearRDSCacheOnDelegateUpdate(t *testing.T) {
 		t.Fatalf("rds cache was cleared by irrelevant delegate virtual service update")
 	}
 }
+
+
+func TestExtractNamespaceForKubernetesService(t *testing.T) {
+	tests := []struct {
+		hostname string
+		want     string
+	}{
+		{
+			"foo.ns.svc.cluster.local",
+			"ns",
+		},
+		{
+			"foo.svc.cluster.local",
+			"",
+		},
+		{
+			"svc.ns.svc.svc.svc",
+			"ns",
+		},
+		{
+			".svc.",
+			"",
+		},
+		{
+			"..svc.",
+			"",
+		},
+		{
+			"x.svc.",
+			"",
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.hostname, func(t *testing.T) {
+			got, err := extractNamespaceForKubernetesService(tt.hostname)
+			if (err != nil) != (tt.want == "") {
+				t.Fatalf("unexpected error = %v", err)
+			}
+			if got != tt.want {
+				t.Fatalf("got = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
